First map on skills.
An inital wardly map based on my skill sets.

There are additional skills to add but need to view this without causing too much noise.
This will form part of a future blog post.

There are no pipelines or inertia at this stage because that require further thought and context as to what that means in relation to the skills narrative.